Provide cmake option to disable testing
CUnit does not need to be a dependency for installation and usage, so
this patch adds a cmake option (WITH_TESTS) that can be used to disable
building of the test suite. WITH_TESTS is set to TRUE by default to keep
the default behavior consistent.

This patch also adds a one-liner to the README mentioning the option.